+++ CHANGELOG.md
@@ -4,15 +4,21 @@
 
 - Fixing deps module, golang fn: checksum added
 - CircleCI mdule defines required jq
+- Fixing PATH issue for binary deps
 
 ### Added
 
 - uaadb start added
 - identity server start added
+- `make dev` added to mac based development
+- `pull` command added
+- `logs` command added
 
 ### Removed
 
 ### Changed
+
+- Docker containers are managed by **docker-compose**
 
 ## [v0.0.7] - 2015-03-26